add signable-bytes subcommand
This subcommand accepts JSON input from a file or stdin, and
emits the appropriate signable bytes for the given transaction to
stdout. Pipes and redirection to files work.

We do need to specify which transaction is intended, because it is
impossible to determine this information from simple inspection of the
JSON.