More println debugging
Emitting the tx hash here allows us to grep for the tx hash in the
ndaunode logs. Adding the address to the relevant ndaunode log allows
us to verify that the correct address is being used, and look more
deeply into what was going on with the relevant account.

It turns out that this is all caused by correct behavior of the transfer,
because the accounts all had the default settlement period of 2 days,
and had no available balance.