feature: convert an API key and secret into a time-limited access token
Features of the bitmart api:

- sending JSON to the API reliably causes 500 internal server error
- encryption doesn't operate on bytes, but on ascii of lowercase hex of bytes

This stuff works despite those features, but it took all day to figure
them out.

Ended up with a python implementation based on their example (as always,
their example doesn't actually work straight out of the box), because it was
quick to iterate on. Once that worked, translated the go code from there.