don't lowercase memo field when computing client secret
The client secret is a hash, so very naturally it's sensitive to changes
in capitalization. The key used in the hash is a composite of three fields,
two of which are hex strings which must always be lowercase letters and digits.

The bug was that the client secret calculation used to lowercase the entire
composite key. The problem was that this also lowercased the third field, which
does _not_ necessarily need to be all-lowercase.

The solution was to lowercase only the first two fields, independently, not the
entire composite key field.