Finish test.sh
This script now returns an appropriate exit code based on whether or not a
transaction was generated during its execution.

Note that it will falsely report success if other transactions come through
while it runs. It only works when run against a net on which you can be sure
that no external transactions will come in.