Separate docker build into logical stages
This is intended to substantially improve caching behavior. The
drawback of this setup is that it has become harder to manually
replicate the exact process; it's more important than it used
to be to use docker/bin/buildimage.sh.

For the speed increases we're seeing here, that should be an
acceptable tradeoff.