remove websocket connection to tm
That was a good, elegant idea, but it just doesn't work with our
existing deploy infrastructure, and it's really not worth
reengineering our whole node structure for this. We can still
ensure that the OTSs are updated after a sale, it's just a bit
more synchronous/blocky this way.

That's why we change the send method, incidentally: using
SendCommit ensures the tx is fully on the blockchain bfore
unblocking, meaning that it is definitely safe to query the
total issue count in `updateOTSs` immediately thereafter.