rewrite runcontainer.sh in python
At almost 400 lines of code, much of it devoted to input parsing,
runcontainer.sh was at the limit of maintainability. I needed to
add features to it to ensure an appropriately flexible port mapping
for postgres, and it just wasn't worth it to try to extend the
shell script anymore. Rewriting in Python took a little while,
but should be much more maintainable and extensible in the future.

This commit includes the necessary improvements for postgres support,
and dramatically simplifies the usage.