generate and apply node user password
This password is in plaintext to anyone who can ssh into the
container as postgres or root; it's not exactly high security.
On the other hand, so is the databsae master password and a
bunch of other secrets; there's not a ton we can do to change
that.

The nice thing about doing it this way is that it's available
externally if we need it, but it appears nowhere in our source
repos. Simultaneously, it's easy for ndaunode to read and use.

It's a compromise, but it should be good enough.