Add --sum option to query scripts
The forAllAccounts.py and forAllTransactions.py offer a --count option to return only the number of records retrieved. This similar options requires a field specifier, and only the sum of the amounts in that field is returned. No type checking is performed to determine whether the requested field is numeric - not worth the trouble.